Added a portfolio manager application
The portfolio manager receives a request message with an account number in the payload. If previously unseen, will generate a random portfolio and save that for future requests for the same account number. If a request for a known account number, will return that from memory.
All portfolio state resets upon restart.